site/backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-storages
boto3
Pillow
python-dotenv
psycopg2-binary
dj-database-url
gunicorn
whitenoise
sentry-sdk
edge-tts==7.2.7
aiohttp==3.13.3